requirements.txt
scikit-learn
kubernetes
numpy
pandas
xgboost
kafka-python
protobuf==3.20.1
